Cover art
Floral Still Life by Hans Bollongier
PD status: CC0 grant from Rijksmuseum on http://hdl.handle.net/10934/RM0001.COLLECT.6117 (click on Object Data / Public domain)

Unfortunately no black tulip paintings exist, but apparently the tulip in the centre of this painting is the most expensive bulb sold at the height of tulip mania, which seems appropriate for this book. As the title covers the vase in all crops I’ve gone for a tight one to spray the entire cover with a flood of tulips.